Ctrl-L (Clear) blanks the current window.
Problems:
* Bug#709811	https://bugzilla.gnome.org/show_bug.cgi?id=709811
* Bug#550885	https://bugzilla.gnome.org/show_bug.cgi?id=550885

Will Thompson (last bug report) has proposed to design the clearance of the
chat window like in a linux terminal. I think, thats a good idea and it
solves also the upper problem (that some parts of the history are deleted
and older ones, which were not displayed before the clearance command, show
up by scrolling up)

I changed empathy_theme_adium_clear() so that the Javascript function
'clearPage()' is exectued instead of re-loading the template. The
Javascipt function is implemented in the data/Template html file.