Probably fix icons in search bar
The previous/next icons have been missing for a while, and I finally
decided I cared to fix them. Also, use a better close icon while we're
at it.

This is untested because I cannot figure out how to actually run
empathy-chat in jhbuild. 'jhbuild run empathy' seems to work for the
roster, but start the system version of empathy-chat. Oh well.