call: fix the video preview
The port to clutter-gst 3.0 broke the video preview window.
The rounded texture effect is now provided by a ClutterEffect
sub-class, and the horizontal flip of the preview is done
directly with a cogl transformation. This last modification
using cogl_push_matrix()/cogl_pop_matrix(), the deprecation
warning of cogl has been lowered from 1.14 to 1.8.

https://bugzilla.gnome.org/show_bug.cgi?id=751185