tests: Actually test that hostnames of pinned certificates are verified
This test case is about ensuring that a pinned certificate won't be
validated if the wrong hostname is used.

If we don't add the pinned certificate to our database, then checks for
pinning are going to fail regardless of the hostname being used. The
correct certificate-hostname pair needs to be in the database to ensure
that the hostnames are being matched as advertised.

https://bugzilla.gnome.org/show_bug.cgi?id=780160